+++ db/migrations/000093_rent_pricing_flags.down.sql
@@ -1,5 +1,5 @@
 ALTER TABLE rent
-    MODIFY price DECIMAL(10, 2) NOT NULL,
+    MODIFY price BIGINT NOT NULL,
     DROP COLUMN price_to,
     DROP COLUMN negotiable,
     DROP COLUMN hide_phone;
